services/embedder/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
torch==2.3.1+cu121 --extra-index-url https://download.pytorch.org/whl/cu121
torchvision==0.18.1+cu121 --extra-index-url https://download.pytorch.org/whl/cu121
pillow==10.4.0
numpy==1.26.4
requests==2.32.3
accelerate==0.34.2
